requirements.txt
Flask
Pillow
numpy
scikit-learn
werkzeug